Shorten hint names
