+++ tests/Immediate.Handlers.Tests/GeneratorTests/Behaviors/SingleBehaviorTest.SingleBehavior_assemblies=Msdi#Dummy.GetUsersQuery.g.verified.cs
@@ -1,4 +1,4 @@
-//HintName: Immediate.Handlers.Templates.Dummy.GetUsersQuery.g.cs
+﻿//HintName: Dummy.GetUsersQuery.g.cs
 using Microsoft.Extensions.DependencyInjection;
 
 #pragma warning disable CS1591
