+++ tests/Immediate.Handlers.Tests/GeneratorTests/Behaviors/SingleBehaviorTest.SingleBehavior_assemblies=Msdi#ServiceCollectionExtensions.g.verified.cs
@@ -1,4 +1,4 @@
-//HintName: Immediate.Handlers.ServiceCollectionExtensions.g.cs
+﻿//HintName: ServiceCollectionExtensions.g.cs
 #pragma warning disable CS1591
 
 namespace Microsoft.Extensions.DependencyInjection;
@@ -9,7 +9,6 @@ public static class HandlerServiceCollectionExtensions
 		this global::Microsoft.Extensions.DependencyInjection.IServiceCollection services)
 	{
 		services.AddScoped(typeof(global::Dummy.LoggingBehavior<,>));
-		services.AddScoped(typeof(global::Dummy.SecondLoggingBehavior<,>));
 		global::Dummy.GetUsersQuery.AddHandlers(services);
 		
 		return services;
